requirements.txt
selenium
pytest
pytest-rerunfailures
pytest-html
pdoc3
allure-pytest
python-dotenv
